malware/RatMuseum-master/Moderat/requirements.txt
coloredlogs
twisted
pillow
matplotlib
qt4reactor
pycrypto
selenium
requests
pyqode.python
pyaudio